+++ examples/resources/rubrik_aws_cnp_account_attachments/import-by-identity.tf
@@ -1,0 +1,6 @@
+import {
+  to = rubrik_aws_cnp_account_attachments.attachments
+  identity = {
+    id = "51ecb385-e4a5-410d-8604-50170378b7a0"
+  }
+}