+++ internal/provider/framework_resource_aws_cnp_account_attachments_move.go
@@ -1,0 +1,115 @@
+// Copyright 2026 Rubrik, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to
+// deal in the Software without restriction, including without limitation the
+// rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
+// sell copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+// DEALINGS IN THE SOFTWARE.
+
+package provider
+
+import (
+	"context"
+	"strings"
+
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
+)
+
+func (r *awsCnpAccountAttachmentsResource) MoveState(ctx context.Context) []resource.StateMover {
+	tflog.Trace(ctx, "awsCnpAccountAttachmentsResource.MoveState")
+
+	return []resource.StateMover{
+		r.moveStateV0(),
+	}
+}
+
+// moveStateV0 moves v0 state from the polaris_aws_cnp_account_attachments
+// resource in the rubrikinc/polaris provider to the
+// rubrik_aws_cnp_account_attachments resource.
+func (r *awsCnpAccountAttachmentsResource) moveStateV0() resource.StateMover {
+	return resource.StateMover{
+		SourceSchema: &schema.Schema{
+			Attributes: map[string]schema.Attribute{
+				keyID: schema.StringAttribute{
+					Computed: true,
+				},
+				keyAccountID: schema.StringAttribute{
+					Required: true,
+				},
+				keyFeatures: schema.SetAttribute{
+					ElementType: types.StringType,
+					Required:    true,
+				},
+				keyRoleChainingAccountID: schema.StringAttribute{
+					Optional: true,
+				},
+			},
+			Blocks: map[string]schema.Block{
+				keyInstanceProfile: schema.SetNestedBlock{
+					NestedObject: schema.NestedBlockObject{
+						Attributes: map[string]schema.Attribute{
+							keyKey: schema.StringAttribute{
+								Required: true,
+							},
+							keyName: schema.StringAttribute{
+								Required: true,
+							},
+						},
+					},
+				},
+				keyRole: schema.SetNestedBlock{
+					NestedObject: schema.NestedBlockObject{
+						Attributes: map[string]schema.Attribute{
+							keyKey: schema.StringAttribute{
+								Required: true,
+							},
+							keyARN: schema.StringAttribute{
+								Required: true,
+							},
+							keyPermissions: schema.StringAttribute{
+								Optional: true,
+							},
+						},
+					},
+				},
+			},
+		},
+		StateMover: func(ctx context.Context, req resource.MoveStateRequest, res *resource.MoveStateResponse) {
+			tflog.Trace(ctx, "awsCnpAccountAttachmentsResource.moveStateV0")
+
+			if !strings.HasSuffix(req.SourceProviderAddress, "rubrikinc/polaris") &&
+				!strings.HasSuffix(req.SourceProviderAddress, "rubrikinc/rubrik") {
+				return
+			}
+			if req.SourceTypeName != keyPolaris+"_"+keyAWSCNPAccountAttachments {
+				return
+			}
+			if req.SourceSchemaVersion != 0 {
+				return
+			}
+
+			var state awsCnpAccountAttachmentsModel
+			res.Diagnostics.Append(req.SourceState.Get(ctx, &state)...)
+			if res.Diagnostics.HasError() {
+				return
+			}
+
+			res.Diagnostics.Append(res.TargetState.Set(ctx, &state)...)
+		},
+	}
+}